Intern phrase pieces
Now once a Phrase is concatenated into another Phrase, it
is marked as interned.  Any attempt to cat to it will spill
it into a clone.

We'll do what we can to cheaply respect that immutability. I
don't want to hide the Phrase pieces on a metatable since this
will be the hottest component of Orb with the possible exception
of the Node constructor inside Grammar.