Add __DEFAULT rule option for metatables
This is useful when defining a base class that can provide error
stubs for unimplemented methods, to give more useful feedback.

It's also annoying when defining metas for grammar categories,
then forgetting to add them to the table, to get obscure nil
errors.

I'm not convinced this is the best way to do it, however;
magic parameters are kind of messy.