Corrects offset behavior in Grammar, Peg
That was easier than I feared, and only touched Orb in one place,
which was easy to remove since it's doing the wrong thing.

The existing Orb parser is deprecated code, these recent changes
being intended to make a better parser possible, so I'm not going
to try and fix what we have.