Remove redundant offset setting in make_ast_node
Also removes all the prose around fixing the offset.

That was crucial to making the fix, but doesn't really belong in
the file once it was done.  This is a hidden advantage of literate
programming; although in this case I wrote the fix as fast as I
worked out the solution, in a more complex case the reasoning
could have sat there for as long as it was still needed.

Yes, it would be *possible* to just put it in a comment. But it
would be ugly and unsatisfying, and less likely to happen, without
the availability of markup.

Also, we aren't doing anything with keywords like #Todo yet, but
it won't be long before we do, allowing for situational awareness
of tasks and their state of completion from `br orb agenda`.