Avoid substring allocation if finish not provided
I'm not sure about this change, because it complicates the logic,
but it does seem to work correctly and it does avoid allocation if
only a start position is offered.

Thing is, I'm not convinced this situation will arise very often.

And if it does, since the grammar will parse to the end of the
string (if it has the opportunity), I wonder if we really save
much here.

Ah well. I'm trying to make Grammar efficient, because parsing
algorithms do a lot of hot looping.  So I'll leave it in and hope
that I got it right.