Better semantics for $name operations
Since the match determines the identity of the string, we match on
length instead.  This allows us to use the same syntax for Orb
code blocks as we use for Lua long strings, text styling, and other
common patterns.

It may be that we want a way to express an exact match, an example
might be begin and end blocks in Ada.  We can come up with a way
to handle that at another time.

I also think semantically this does a better job of generalizing
the `repeats` operator, since `rule$2` matches exactly two of
`rule`.