Only add functions if used in grammar
It's a small gesture, and anyone's guess if it's more efficient to
build five functions every time, or to repeatedly walk the grammar
tree.

I prefer it this way. A case could be made that I should build the
rules in Lua, instead of out of strings, and make a fenv for the
Grammar; but then I'd have to either stick Peg stuff in the Grammar
module (which feels a bit off) or sacrifice the compatibility with
stock Lua, which would seem a shame.