Remove NotThis predicate
There's no difference between `rule !not-follow` and
`rule - not-follow`, and we were already violating the general
principle that e.g. `rule/other` is the same as `rule / other`.

Combine this with the bad parsing of `rule -1`, which was reducing
to `V"rule" - P(1)` instead of `V"rule" * P(-1)`, and we're better
off just dumping the rule.  This lets us keep the precedence of
rule matches we already have, and makes -1 match the end of a
string without needing to put parentheses around it.

It was the only case of having a synonym for a rule, and it was a
weird place to put it.  Lpeg has two ways to say the same thing
since the alternative is to define a unary minus but not a binary
one, and that's weird.  We don't have such a justification.