Corrects linePos()
It's no longer buggy as hell and returns both node.first and
node.last.

It's tempting to memoize the array of start, end values for each
line, like the node:lines() function does (although now that's even
sillier).

But I'm going to hold off on doing anything which would mean having
to cache-bust if the in-memory Node is mutated.