Fix bug when final line has no newline
I'm surprised this didn't cause more problems, but the units we
work with in Orb tend to have line ends.

That said, ending a Doc without a newline is definitely valid, and
the logic was also simply incorrect: it would have failed badly on
any grammar which has a "dent" rule, for newline + indentation, if
trying to get a linePos from the preceding Node.