Add Node:selectFrom(pred, index)
Link resolution in Doc requires finding the next match on a
predicate after the place where the obelus is found.

So we introduce `selectFrom` to handle this use case.  `selectTo`
is an obvious pair, but we'll add it when we need it.

`selectInRange` seems pretty pointless, since any reasonable range
will live inside a Node, which can simply be used to perform the
selection.