Implement node:replace(branch, index)
It's the same algorithm as insertion, so I left logic for
replacement and went back to fill it in.

Crude though it is, the sessions are already saving me time: I
put the replace logic where the (not replace) logic belonged and
it broke my sessions immediately, saving me having to do the same
thing manually at the repl, with the mental shift from coding to
repling which that entails.