Complete, untested TOML grammar, minus dates
I'll just go ahead and take the opportunity to make a date-parsing
PEG and append it.

I would be very pleased if this works, unlikely though it feels to
write an entire grammar without any testing.

But translating to PEG from ABNF is not exceptionally difficult,
barring the fact that we need to order some things differently.