Eliminate leading false patterns + doc pass
The only semantic change here is to remove leading falsy patterns
from R and S, since a moderately more complex compiling step is to
be preferred to a runtime penalty.

The bulk of the changes are to the documentation, adding a bit and
bringing the rest in line with prevailing style.

Which still has no guide, and I should really do something about
that.