Add node:selectBack()
This is equivalent to node:selectFrom(index), except it does what
`:selectFrom` should, and derives the index from the Node.

As I think about it, the index should actually be an optional
parameter in both cases. But I'll have to grep for uses of the
previous method before I can mess with it.