Transforms existing Grammar into combinator
This isn't the final form, but it makes the point: the definition
function is independent of how the evaluation environment decides
to act on the information.

Eventually we can inject API-compatible substitutes for the Lpeg
primitives, for hammer and other contemplated uses, but let's not
get ahead of ourselves here.