Dji combinator
Be wary of any error you may encounter in running this algorithm.

I have merely named it, let alone sketched it, told tall tales
about it, proved it correct, or, heaven forfend, implemented it.

Caveats belong to the lector, and do not, do not, forget to tip
your waitress.

Here all week, folks.