Curry the dji combinator to the recognizer
Which isn't correct...? yet. But is a gesture in the direction of
correctness.  Which is all that we could as for from the first
commit of the intercalary week.

Now, if you will excuse me, some dust has fallen upon my
microphone, unaccountably, and I must needs dust it less I should
inhale some without strictly meaning to.

Ahem. Ahem I say. There shall be no snickering in this
establishment!  A decent place it is, the watering hole, and a
decent place it shall remain!

seriously thanks for tipping your waitress it's rough out there
with the coof and all. tough crowd but we got our reasons. moving on.