Base combining rule
A rule must be at least as powerful as its least powerful
subrule, so this correctly covers a number of cases.

Next step is to assure that it reports everything as at least
regular, which will cover that power level, then we generalize our
recursion analysis from the subrule thinsg to get the rules which
recurse.