Dramatically simpler parser
This breaks all the early steps in the synthesis chain but this was
part of the point of using a derived synth node, changes to the
parser mean changes to the coupling, rather than pervasive through
they system.

It's much, much faster than either the original or the previous
edition.