Constraints satisfied (results not guaranteed)
This does reach a fixed point, it seems to promote the traits it
needs to onto everything, and I /believe/ the various compound
inferences are broadly correct.

This needs a round of /proof/ though, and a shakedown for bugs.

That said, it does open the possibility of doing the next step of
the analysis.

That's nice. I like that.