Viable Node2
We've got a lot of work ahead of us, but I can create v2 Nodes
and print them (reasonably) compactly.

The next step, after merging this with trunk, is to build out a
spec, and write some sessions as I extend the interface.

Full revision of sessions from helm is in the pipeline, which
I'm looking forward to. Excelsior!