Add Node:gather(pred)
I like this pattern of "find one, put it on the stack, recurse",
and just using `pack` for when we want an array.

Belatedly realizing that I wasn't using it earlier because of the
bad experiences ~@antirez had with multiple return values, but
that was stock Lua: LuaJIT doesn't mind at all.